+++ Mile.Helpers/Mile.Helpers.h
@@ -59,4 +59,18 @@ EXTERN_C HRESULT WINAPI MileSetWindowCaptionColorAttribute(
     _In_ HWND WindowHandle,
     _In_ COLORREF Value);
 
+/**
+ * @brief Allows the window frame for this window to be drawn in dark mode
+ *        colors when the dark mode system setting is enabled.
+ * @param WindowHandle The handle to the window for which the attribute value
+ *                     is to be set.
+ * @param Value TRUE to honor dark mode for the window, FALSE to always use
+ *              light mode.
+ * @return If the function succeeds, it returns S_OK. Otherwise, it returns an
+ *         HRESULT error code.
+*/
+EXTERN_C HRESULT WINAPI MileSetWindowUseImmersiveDarkModeAttribute(
+    _In_ HWND WindowHandle,
+    _In_ BOOL Value);
+
 #endif // !MILE_WINDOWS_HELPERS